% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ak1991}
\alias{ak1991}
\title{Census data for Angrist and Krueger (1991).}
\format{
A tibble with 1,063,634 rows and 6 columns:
\describe{
\item{log_wage}{Log weekly wage.}
\item{education}{Years of education.}
\item{quarter_birth}{Quarter of birth.}
\item{birthyear}{Indicator variable for Hispanic.}
\item{census}{Year of census observation comes from.}
\item{race}{Respondent race (white or black).}
}
}
\source{
Angrist Data Archive.
}
\usage{
ak1991
}
\description{
Census data for Angrist and Krueger (1991).
}
\keyword{datasets}
