% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capm}
\alias{capm}
\title{Estimates from Capital Asset Pricing Model for various stocks.}
\format{
A tibble with 1,096 rows and 3 columns:
\describe{
\item{   ticker}{Stock ticker.}
\item{   parameter}{Alpha and beta estimates.}
\item{   estimate}{Point estimate.}
}
}
\source{
Estimated from data used in Matt Taddy's Business Data Science
}
\usage{
capm
}
\description{
Data used for class as an example of pivot_wider.
}
\keyword{datasets}
