% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_2016}
\alias{election_2016}
\title{Electoral college results from 2016 U.S. Presidential Election.}
\format{
A tibble with 51 rows and 3 columns:
\describe{
\item{   state}{State including D.C.}
\item{   electoral_votes}{Number of electoral votes by state.}
\item{   winner}{Winning candidate by state.}
}
}
\usage{
election_2016
}
\description{
Electoral college results from 2016 U.S. Presidential Election.
}
\keyword{datasets}
