% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expenditure_data1}
\alias{expenditure_data1}
\title{Expenditure data from a hypothetical household survey.}
\format{
A tibble with 100 rows and 4 columns:
\describe{
\item{   id}{Respondent id.}
\item{   food}{Expenditure on food.}
\item{   clothing}{Expenditure on clothing.}
\item{   housing}{Expenditure on housing.}
\item{   alcohol}{Expenditure on alcohol.}
}
}
\source{
Created by Andrew S. Griffen
}
\usage{
expenditure_data1
}
\description{
Data used for class.
}
\keyword{datasets}
