% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expenditure_data2}
\alias{expenditure_data2}
\title{Expenditure data from a hypothetical household survey.}
\format{
A tibble with 100 rows and 201 columns:
\describe{
\item{   id}{Respondent id.}
\item{   item"i"}{Expenditure on item i.}
}
}
\source{
Created by Andrew S. Griffen
}
\usage{
expenditure_data2
}
\description{
Data used for class.
}
\keyword{datasets}
