% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gettysburg}
\alias{gettysburg}
\title{Text of the Gettysburg Address.}
\format{
A character.
}
\source{
President Abraham Lincoln, 16th President of the United States of America.
}
\usage{
gettysburg
}
\description{
Famous speech about America's national purpose.
}
\keyword{datasets}
