% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griffendata-package.R
\docType{package}
\name{griffendata-package}
\alias{griffendata}
\alias{griffendata-package}
\title{griffendata: Datasets for R for Empirical Economics Research}
\description{
Example datasets used for my course "R for Empirical Economics Research".
}
\author{
\strong{Maintainer}: Andrew Griffen \email{andrew.griffen@gmail.com}

}
\keyword{internal}
