% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mascots}
\alias{mascots}
\title{Mascot data from the 2023 Yuru-Kyara Grand Prix "Yuruverse 2023"}
\format{
A tibble with 41 rows and 6 columns:
\describe{
\item{   name}{Mascot name}
\item{   prefecture}{Japanese prefecture}
\item{   prefecture_en}{Japanese prefecture (EN)}
\item{   region}{Region of Japan}
\item{   place}{Place in the Grand Prix}
\item{   group}{Mascot sponsor}
\item{   description}{Description of mascot}
}
}
\source{
\url{https://www.yurugp.jp}
}
\usage{
mascots
}
\description{
Mascot data from the 2023 Yuru-Kyara Grand Prix "Yuruverse 2023"
}
\keyword{datasets}
