% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oecd}
\alias{oecd}
\title{Time series data on life expectancy for OECD countries.}
\format{
A tibble with 100 rows and 201 columns:
\describe{
\item{   country}{Country.}
\item{   year}{Year.}
\item{   sex}{Sex.}
\item{   life_expectancy}{Life expectancy in years.}
}
}
\source{
Organisation for Economic Co-operation and Development (OECD) data
}
\usage{
oecd
}
\description{
Data used for class.
}
\keyword{datasets}
