% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow_deaths}
\alias{snow_deaths}
\title{Location and number of cholera deaths in the 1854 outbreak in Soho, London.}
\format{
A tibble with 250 rows and 4 columns:
\describe{
\item{   lng}{longitude of address}
\item{   lat}{longitude of address}
\item{   deaths}{number of deaths at the address}
}
}
\source{
\url{http://blog.rtwilson.com/?s=snow}
}
\usage{
snow_deaths
}
\description{
Location and number of cholera deaths in the 1854 outbreak in Soho, London.
}
\keyword{datasets}
