% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow_pumps}
\alias{snow_pumps}
\title{Location and number of water pumps in 1854 Soho, London.}
\format{
A tibble with 8 rows and 2 columns:
\describe{
\item{   lng}{longitude of pump}
\item{   lat}{longitude of pump}
}
}
\source{
\url{http://blog.rtwilson.com/?s=snow}
}
\usage{
snow_pumps
}
\description{
Location and number of water pumps in 1854 Soho, London.
}
\keyword{datasets}
